refactor, move ManyComponentWriter class into the main aspect, move more legacy code into the aspect
