+++ scopes/component/merging/index.ts
@@ -1,7 +1,7 @@
 import { MergingAspect } from './merging.aspect';
 
-export { mergeReport } from './merge-cmd';
-export type { MergingMain, ComponentMergeStatus } from './merging.main.runtime';
+export { mergeReport, applyVersionReport, conflictSummaryReport } from './merge-cmd';
+export type { MergingMain, ComponentMergeStatus, ApplyVersionResults } from './merging.main.runtime';
 export { ConfigMergeResult } from './config-merge-result';
 export default MergingAspect;
 export { MergingAspect };
