+++ scopes/component/merging/merge-cmd.ts
@@ -3,14 +3,13 @@ import { Command, CommandOptions } from '@teambit/cli';
 import { compact } from 'lodash';
 import { WILDCARD_HELP, AUTO_SNAPPED_MSG, MergeConfigFilename } from '@teambit/legacy/dist/constants';
 import {
-  ApplyVersionResults,
-  conflictSummaryReport,
   getMergeStrategy,
-  applyVersionReport,
+  FileStatus,
+  ApplyVersionResult,
 } from '@teambit/legacy/dist/consumer/versions-ops/merge-version';
 import { isFeatureEnabled, BUILD_ON_CI } from '@teambit/legacy/dist/api/consumer/lib/feature-toggle';
 import { BitError } from '@teambit/bit-error';
-import { MergingMain } from './merging.main.runtime';
+import { ApplyVersionResults, MergingMain } from './merging.main.runtime';
 import { ConfigMergeResult } from './config-merge-result';
 
 export class MergeCmd implements Command {
@@ -217,4 +216,45 @@ ${mergeSnapError.message}
     getConflictSummary() +
     getSummary()
   );
+}
+
+export function applyVersionReport(components: ApplyVersionResult[], addName = true, showVersion = false): string {
+  const tab = addName ? '\t' : '';
+  return components
+    .map((component: ApplyVersionResult) => {
+      const name = showVersion ? component.id.toString() : component.id.toStringWithoutVersion();
+      const files = Object.keys(component.filesStatus)
+        .map((file) => {
+          const note =
+            component.filesStatus[file] === FileStatus.manual
+              ? chalk.white(
+                  'automatic merge failed. please fix conflicts manually and then run "bit install" and "bit compile"'
+                )
+              : '';
+          return `${tab}${component.filesStatus[file]} ${chalk.bold(file)} ${note}`;
+        })
+        .join('\n');
+      return `${addName ? name : ''}\n${chalk.cyan(files)}`;
+    })
+    .join('\n\n');
+}
+
+export function conflictSummaryReport(components: ApplyVersionResult[]): string {
+  const tab = '\t';
+  return compact(
+    components.map((component: ApplyVersionResult) => {
+      const name = component.id.toStringWithoutVersion();
+      const files = compact(
+        Object.keys(component.filesStatus).map((file) => {
+          if (component.filesStatus[file] === FileStatus.manual) {
+            return `${tab}${component.filesStatus[file]} ${chalk.bold(file)}`;
+          }
+          return null;
+        })
+      );
+      if (!files.length) return null;
+
+      return `${name}\n${chalk.cyan(files.join('\n'))}`;
+    })
+  ).join('\n');
 }
