+++ src/consumer/versions-ops/merge-version/index.ts
@@ -1,10 +1,3 @@
 export { default as threeWayMerge } from './three-way-merge';
-export { FileStatus, getMergeStrategyInteractive, getMergeStrategy, MergeOptions } from './merge-version';
-export {
-  MergeStrategy,
-  ApplyVersionResults,
-  ApplyVersionResult,
-  FailedComponents,
-  applyVersionReport,
-  conflictSummaryReport,
-} from './merge-version';
+export { FileStatus, FilesStatus, getMergeStrategyInteractive, getMergeStrategy, MergeOptions } from './merge-version';
+export { MergeStrategy, FailedComponents, ApplyVersionResult } from './merge-version';
