+++ src/consumer/versions-ops/merge-version/merge-version.ts
@@ -1,10 +1,7 @@
 import chalk from 'chalk';
-import { compact } from 'lodash';
 import { BitId } from '../../../bit-id';
 import GeneralError from '../../../error/general-error';
 import { resolveConflictPrompt } from '../../../prompts';
-import { AutoTagResult } from '../../../scope/component-ops/auto-tag';
-import Component from '../../component/consumer-component';
 
 export const mergeOptionsCli = { o: 'ours', t: 'theirs', m: 'manual' };
 export const MergeOptions = { ours: 'ours', theirs: 'theirs', manual: 'manual' };
@@ -23,21 +20,9 @@ export const FileStatus = {
 };
 // fileName is PathLinux. TS doesn't let anything else in the keys other than string and number
 export type FilesStatus = { [fileName: string]: keyof typeof FileStatus };
-export type ApplyVersionResult = { id: BitId; filesStatus: FilesStatus };
 export type FailedComponents = { id: BitId; failureMessage: string; unchangedLegitimately?: boolean };
-export type ApplyVersionResults = {
-  components?: ApplyVersionResult[];
-  version?: string;
-  failedComponents?: FailedComponents[];
-  resolvedComponents?: Component[]; // relevant for bit merge --resolve
-  abortedComponents?: ApplyVersionResult[]; // relevant for bit merge --abort
-  mergeSnapResults?: { snappedComponents: Component[]; autoSnappedResults: AutoTagResult[] } | null;
-  mergeSnapError?: Error;
-  leftUnresolvedConflicts?: boolean;
-  verbose?: boolean;
-  newFromLane?: string[];
-  newFromLaneAdded?: boolean;
-};
+
+export type ApplyVersionResult = { id: BitId; filesStatus: FilesStatus };
 
 export async function getMergeStrategyInteractive(): Promise<MergeStrategy> {
   try {
@@ -58,45 +43,4 @@ export function getMergeStrategy(ours: boolean, theirs: boolean, manual: boolean
   if (theirs) return MergeOptions.theirs as any;
   if (manual) return MergeOptions.manual as any;
   return null;
-}
-
-export const applyVersionReport = (components: ApplyVersionResult[], addName = true, showVersion = false): string => {
-  const tab = addName ? '\t' : '';
-  return components
-    .map((component: ApplyVersionResult) => {
-      const name = showVersion ? component.id.toString() : component.id.toStringWithoutVersion();
-      const files = Object.keys(component.filesStatus)
-        .map((file) => {
-          const note =
-            component.filesStatus[file] === FileStatus.manual
-              ? chalk.white(
-                  'automatic merge failed. please fix conflicts manually and then run "bit install" and "bit compile"'
-                )
-              : '';
-          return `${tab}${component.filesStatus[file]} ${chalk.bold(file)} ${note}`;
-        })
-        .join('\n');
-      return `${addName ? name : ''}\n${chalk.cyan(files)}`;
-    })
-    .join('\n\n');
-};
-
-export function conflictSummaryReport(components: ApplyVersionResult[]): string {
-  const tab = '\t';
-  return compact(
-    components.map((component: ApplyVersionResult) => {
-      const name = component.id.toStringWithoutVersion();
-      const files = compact(
-        Object.keys(component.filesStatus).map((file) => {
-          if (component.filesStatus[file] === FileStatus.manual) {
-            return `${tab}${component.filesStatus[file]} ${chalk.bold(file)}`;
-          }
-          return null;
-        })
-      );
-      if (!files.length) return null;
-
-      return `${name}\n${chalk.cyan(files.join('\n'))}`;
-    })
-  ).join('\n');
 }
